Device hashmap datastructure
For storing shear history between GPU kernel force-calculations. Need to
write hooks to build from/resync to cpu datastructures in listgranhistory.

Simple standalone test for hashmap is in play/ directory.
$ nvcc -arch sm_13 test_hashmap.cu -o test_hashmap
$ ./test_hashmap

should initialize a cpu hashmap, copy it over to the device, increment each
shear vector by {100,200,300} on the device, and copy it back to the host for
printing.