Symmetric update for particle j working.
On the CPU the same iteration of the inner compute loop is used to
update particle i due to particle j as well as
update particle j due to particle i.

On the GPU we were only considering the first case. Now we do both cases but
separate out the computation into separate threads to avoid data races.

Note: could feasibly update particle j at the same time as i if j is in the same
cell block as i? The thing to ensure is that there are no multiple writers to
the same force/torque/shear location.